requirements.txt
asgiref==3.7.2

cloudinary==1.37.0

coverage==7.4.0

dj-database-url==2.1.0

django-cors-headers==4.3.1



Django==5.0.1

django-filter==23.5


djangorestframework==3.14.0

djangorestframework-simplejwt==5.3.1

drf-spectacular==0.27.0

dj3-cloudinary-storage==0.0.6

gunicorn==21.2.0

pillow==10.1.0

psycopg2-binary==2.9.9

python-decouple==3.8

pytz==2023.3

requests==2.26.0

sqlparse==0.4.2

urllib3==1.26.9